requirements.txt
tensorflow-cpu==2.6.0
fastapi
io
cv2
uvicorn
python-multipart
pillow
matplotlib
pandas
numpy